==> notebooks/01_load_svi_data.ipynb <==
import pandas as pd
import os



svi_csv_path = r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\raw\SVI2022_US_COUNTY.csv"

svi_df = pd.read_csv(
    svi_csv_path,
    dtype={"FIPS": str},
    encoding="latin1",
    low_memory=False
)

svi_df.head()


svi_df.shape

# Ensure FIPS is a 5-character string (leading zeros preserved)
svi_df["FIPS"] = svi_df["FIPS"].astype(str).str.zfill(5)



important_cols = [
    "FIPS", "STATE", "ST_ABBR", "COUNTY", "E_TOTPOP",

    # Percent-based vulnerability indicators (EP_ columns)
    "EP_POV150", "EP_UNEMP", "EP_NOHSDP", "EP_SNGPNT",
    "EP_MINRTY", "EP_LIMENG", "EP_UNINSUR", "EP_DISABL", "EP_NOVEH",

    # Composite vulnerability index scores
    "RPL_THEME1", "RPL_THEME2", "RPL_THEME3", "RPL_THEME4", "RPL_THEMES"
]


svi_clean = svi_df[important_cols].copy()

svi_clean.head()

#save cleaned data to new csv file
svi_clean.to_csv(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\processed\svi_cleaned.csv", index=False)
