==> notebooks/02_load_places_data.ipynb <==
import pandas as pd

places_df_og = pd.read_csv(r'C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\raw\PLACES__Local_Data_for_Better_Health,_County_Data_2024_release_20251204.csv')

places_df_og.head()



places = places_df_og[places_df_og["Data_Value_Type"] == "Crude Prevalence"]

places_df_og["Data_Value_Type"].unique()


places_df = places_df_og[places_df_og["Data_Value_Type"] == "Crude prevalence"]


places_df.head(20)

#keep only crude prevalence rows
#next remove unneeded columns
places_df = places_df[[
    "StateAbbr",
    "LocationName",
    "Category",
    "Short_Question_Text",
    "Data_Value",
    "Geolocation"
]]


places_df.head()  

places_df["StateAbbr"] = places_df["StateAbbr"].str.upper().str.strip()
places_df["LocationName"] = places_df["LocationName"].str.upper().str.strip()


places_df.head()

fips_url = "https://raw.githubusercontent.com/kjhealy/fips-codes/master/state_and_county_fips_master.csv"
fips_df = pd.read_csv(fips_url)


fips_df.head()

fips_df = fips_df.dropna(subset=["state"])


fips_df["county_clean"] = (
    fips_df["name"]
    .str.replace(" County", "", regex=False)
    .str.upper()
    .str.strip()
)


fips_df

fips_df["FIPS"] = fips_df["fips"].astype(int).astype(str).str.zfill(5)


fips_df.head()

fips_df["state"] = fips_df["state"].str.upper().str.strip()


places_df = places_df.merge(
    fips_df[["FIPS", "state", "county_clean"]],
    left_on=["StateAbbr", "LocationName"],
    right_on=["state", "county_clean"],
    how="left"
)


places_df.head()

places_df = places_df.drop(columns=["state", "county_clean"])


places_df.head()

places_df.to_csv("places_cleaned.csv", index=False)


places_df
