==> notebooks/03_merge_svi_and_places.ipynb <==
import pandas as pd


places = pd.read_csv(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\processed\places_cleaned.csv", dtype={"FIPS": str})
svi = pd.read_csv(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\processed\svi_cleaned.csv", dtype={"FIPS": str})


places

svi

merged_svi_places_df = places.merge(
    svi,
    on="FIPS",
    how="left"   # keep all PLACES rows
)

merged_svi_places_df

merged_svi_places_df.shape

pivot_df = merged_svi_places_df[[
    "FIPS",
    "StateAbbr",
    "STATE",
    "COUNTY",
    "Short_Question_Text",
    "Data_Value",
 
    
    "RPL_THEMES",
    "RPL_THEME1",
    "RPL_THEME2",
    "RPL_THEME3",
    "RPL_THEME4"
]].copy()

pivot_df.head()

wide = pivot_df.pivot_table(
    index=[
        "FIPS",
        "StateAbbr",
        "STATE",
        "COUNTY",
        "RPL_THEMES",
        "RPL_THEME1",
        "RPL_THEME2",
        "RPL_THEME3",
        "RPL_THEME4"
    ],
    columns="Short_Question_Text",
    values="Data_Value"
).reset_index()


wide

# Clean all column names
wide.columns = (
    wide.columns
    .str.strip()
    .str.lower()
    .str.replace(" ", "_")
    .str.replace("-", "_")
    .str.replace("__", "_")
)


wide = wide.rename(columns={
    "fips": "fips",
    "stateabbr": "state",
    "state": "state_full",
    "county": "county_full",
    "rpl_themes": "svi_overall",
    "rpl_theme1": "svi_socioeconomic",
    "rpl_theme2": "svi_household",
    "rpl_theme3": "svi_minority",
    "rpl_theme4": "svi_housing"
})


wide

location_cols = [
    "fips", "state", "state_full", "county_full"
]

svi_cols = [
    "svi_overall", "svi_socioeconomic", "svi_household", 
    "svi_minority", "svi_housing"
]

other_cols = [c for c in wide.columns if c not in location_cols + svi_cols]

wide = wide[location_cols + svi_cols + other_cols]


wide.head() 

[col for col in wide.columns if "diab" in col.lower()]
[col for col in wide.columns if "obes" in col.lower()]
[col for col in wide.columns if "blood" in col.lower()]
[col for col in wide.columns if "heart" in col.lower()]
[col for col in wide.columns if "stroke" in col.lower()]
[col for col in wide.columns if "asth" in col.lower()]


wide.columns.tolist()


#save wide as a cleaned csv file
wide.to_csv(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\processed\svi_places_wide_cleaned.csv", index=False)
