==> notebooks/04_load_cvs_stores.ipynb <==
import pandas as pd

cvs_stores = pd.read_csv(r'C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\raw\store.csv')

cvs_stores.head()

cvs_stores.columns.tolist()


clinic = cvs_stores[cvs_stores["indicatorMinuteClinicService"] == "Y"]

clinic.head()


clinic.columns.tolist()


clinic = clinic[[
    "StoreNumber",
    "minuteClinicID",
    "addressLine",
    "addressState",
    "addressZipCode",
    "geographicLatitudePoint",
    "geographicLongitudePoint"
]].copy()


clinic

clinic = clinic.rename(columns={
    "StoreNumber": "store_id",
    "minuteClinicID": "clinic_id",
    "addressLine": "address",
    "addressState": "state",
    "addressZipCode": "zip",
    "geographicLatitudePoint": "latitude",
    "geographicLongitudePoint": "longitude"
})

# Convert coordinates to numeric
clinic["latitude"] = pd.to_numeric(clinic["latitude"], errors="coerce")
clinic["longitude"] = pd.to_numeric(clinic["longitude"], errors="coerce")

clinic.head()


zip_fips = pd.read_excel(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\raw\ZIP_COUNTY_062025.xlsx")

zip_fips['ZIP'] = zip_fips['ZIP'].astype(str).str.zfill(5)

#merge data
clinic['zip'] = clinic['zip'].astype(str).str.zfill(5)
# merge
merged_clinic = clinic.merge(
    zip_fips[['ZIP', 'COUNTY']],
    left_on='zip',
    right_on='ZIP',
    how='left'
)


merged_clinic

merged_clinic

#save cleaned data 
merged_clinic.to_csv(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\processed\cvs_minute_clinics_cleaned.csv", index=False)
