==> notebooks/05_final_merge.ipynb <==
import pandas as pd


# load cleaned datasets
clinic = pd.read_csv(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\processed\cvs_minute_clinics_cleaned.csv")
svi_places = pd.read_csv(r"C:\Users\14122\OneDrive\Desktop\cvs_heath_project\data\processed\svi_places_wide_cleaned.csv")

# ensure FIPS type matches
clinic['COUNTY'] = clinic['COUNTY'].astype(str).str.zfill(5)
svi_places['COUNTY'] = svi_places['fips'].astype(str).str.zfill(5)

final_df = clinic.merge(
    svi_places,
    on="COUNTY",
    how="left"
)

final_df.head()

clinic_counts = clinic.groupby("COUNTY").size().reset_index(name="clinic_count")
clinic_counts.head()


county_level_df = svi_places.merge(
    clinic_counts,
    on="COUNTY",
    how="left"
)

# fill counties with 0 clinics
county_level_df["clinic_count"] = county_level_df["clinic_count"].fillna(0)




county_level_df.head()

#lookup country 04013
county_level_df[county_level_df['COUNTY'] == '04013']

county_level_df.head()

county_level_df.to_csv("CVS_FINAL_DATASET.csv", index=False)
